item: Refactor Item to use subtypes
It's a slight hack to tell the ECS that it's receiving an Item when in
reality we're giving it a subtype. However, this works just fine because
we give it the proper size to copy over.

This also moves the Item struct to the item.h header to decouple Item
from the component header.