initiative: Add initiative system for variable action cost
Variablbe action cost is a big part of roguelikes, so it's good that we
have it. The Initiative system goes through each entity and increments
its "action points" by the entity-specific increment each turn. When an
entity has >= 0 points, they may make a move. Each move subtracts a
specific amount from the action points. The system is structured
(arbitrarily) so that "one move" costs 100 points.