action: Refactor action component structure and interactions
There is a new model for actions. Each actor has an exclusive
relationship, (HasAction, *), where '*' must be some form of action. On
the actor's turn, this action is unconditionally executed. The action is
*NOT* unset, meaning each AI *MUST* define an action, otherwise the
actor will repeat the last action, which is bad.

This commit also changes MyTurn from a tag that gets added and removed
to a tag the gets enabled and disabled. This prevents unnecessary
archetype migrations. Additionally, actions taken through socket menus
are now respected. This is accomplished in such a way that the player
cannot possible execute two actions on one turn, due to the exclusivity
of the HasAction relationship. Furthermore, MyTurn is automatically
added to any entity that has InitiativeData added or set to it.